requirements.txt
beautifulsoup4==4.12.3
chromedriver_autoinstaller==0.6.4
google_api_python_client==2.130.0
protobuf==5.27.0
pydantic==2.7.2
PyYAML==6.0.1
Requests==2.32.2
selenium==4.21.0
